+++ changelog.md
@@ -1,4 +1,4 @@
-Change Log Key Terms:
+a global variable. Change Log Key Terms:
    added, changed, deprecated, removed, fixed, security
 Optional Terms:
    updated, created, *fix(todo), *debug(todo), *add(todo), *update(todo)
@@ -6,7 +6,7 @@ Optional Terms:
 v1.0
 Name: RPG_Map
 added - map - nested lists
-added - global variables row and col
+added - global variables row and column
 added - contiuous play
 added - "walk" as a main menu option
 added - movement() function
@@ -15,6 +15,22 @@ fixed - movement so player can not walk off "map"
 
 v2.0
 Name: RPG_Map
-*add - change log
+added - change log
 fixed - formatting errors
 fixed - errors in the code
+
+v3.0
+Name : RPG_Inventory
+added - nested dictionaries
+added - empty list
+added - global variable inventory as a list
+added - loop that prints the description of each room
+added - items() function
+added - comments to better explain certain parts of the code
+updated - change log
+updated - version history
+updated - file name
+updated - date the file was submitted
+updated - description of the game
+fixed - formatting errors
+fixed - items so player cannot pick up the same item twice